--- a/pootle/apps/__init__.py
+++ b/pootle/apps/__init__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) Pootle contributors.
+#
+# This file is a part of the Pootle project. It is distributed under the GPL3
+# or later license. See the LICENSE file for a copy of the license and the
+# AUTHORS file for copyright and authorship information.
+
+import importlib
+
+from django.apps import AppConfig
+
+
+class PootleConfig(AppConfig):
+
+    name = "pootle"
+    verbose_name = "Pootle"
+
+    def ready(self):
+        importlib.import_module("pootle.core.receivers")