--- a/pootle/apps/pootle_app/views/admin/dashboard.py
+++ b/pootle/apps/pootle_app/views/admin/dashboard.py
@@ -64,15 +64,15 @@ def server_stats_more(request):
 
         result = {}
         unit_query = Unit.objects.filter(state__gte=TRANSLATED).exclude(
-            store__translation_project__project__code__in=(
+            project__code__in=(
                 'pootle', 'tutorial', 'terminology')
         ).exclude(
-            store__translation_project__language__code='templates').order_by()
+            language__code='templates').order_by()
         result['store_count'] = unit_query.values('store').distinct().count()
         result['project_count'] = unit_query.values(
-            'store__translation_project__project').distinct().count()
+            'project').distinct().count()
         result['language_count'] = unit_query.values(
-            'store__translation_project__language').distinct().count()
+            'language').distinct().count()
         sums = sum_column(unit_query, ('source_wordcount',), count=True)
         result['string_count'] = sums['count']
         result['word_count'] = sums['source_wordcount'] or 0