--- a/pootle/apps/pootle_store/migrations/0014_make_unit_language_notnull.py
+++ b/pootle/apps/pootle_store/migrations/0014_make_unit_language_notnull.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('pootle_store', '0013_denormalize_language'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='unit',
+            name='language',
+            field=models.ForeignKey(related_name='units', to='pootle_language.Language'),
+        ),
+    ]