--- a/pootle/apps/pootle_store/migrations/0017_make_unit_project_notnull.py
+++ b/pootle/apps/pootle_store/migrations/0017_make_unit_project_notnull.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('pootle_store', '0016_denormalize_project'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='unit',
+            name='project',
+            field=models.ForeignKey(related_name='units', to='pootle_project.Project'),
+        ),
+    ]