--- a/pootle/apps/pootle_store/migrations/0019_add_store_name_index.py
+++ b/pootle/apps/pootle_store/migrations/0019_add_store_name_index.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('pootle_store', '0018_index_proj_lang_together'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='store',
+            name='name',
+            field=models.CharField(max_length=128, editable=False, db_index=True),
+        ),
+    ]