--- a/pootle/apps/pootle_store/models.py
+++ b/pootle/apps/pootle_store/models.py
@@ -262,7 +262,7 @@ class UnitManager(models.Manager):
                 store__translation_project__language__code=language_code)
         else:
             units_qs = units_qs.exclude(
-                store__pootle_path__startswith="/templates/")
+                pootle_path__startswith="/templates/")
 
         if project_code:
             units_qs = units_qs.filter(
@@ -282,17 +282,17 @@ class UnitManager(models.Manager):
         if language_code and project_code:
             if filename:
                 return units_qs.filter(
-                    store__pootle_path=pootle_path)
+                    pootle_path=pootle_path)
             else:
                 return units_qs.filter(
-                    store__pootle_path__startswith=pootle_path)
+                    pootle_path__startswith=pootle_path)
         else:
             # we need to use a regex in this case as lang or proj are not
             # set
             if filename:
                 pootle_path = "%s$" % pootle_path
             return units_qs.filter(
-                store__pootle_path__regex=pootle_path)
+                pootle_path__regex=pootle_path)
 
 
 class Unit(models.Model, base.TranslationUnit):