--- a/pootle/apps/pootle_store/models.py
+++ b/pootle/apps/pootle_store/models.py
@@ -318,8 +318,7 @@ class Unit(models.Model, base.TranslationUnit):
     # denormalized TP.project
     project = models.ForeignKey(
         "pootle_project.Project",
-        null=True,
-        blank=True,
+        null=False,
         db_index=True,
         related_name='units')