--- a/pootle/apps/pootle_store/models.py
+++ b/pootle/apps/pootle_store/models.py
@@ -309,6 +309,14 @@ class Unit(models.Model, base.TranslationUnit):
         max_length=255, null=False, unique=False,
         db_index=True, verbose_name=_("Path"))
 
+    # denormalized TP.language
+    language = models.ForeignKey(
+        "pootle_language.Language",
+        null=True,
+        blank=True,
+        db_index=True,
+        related_name='units')
+
     source_f = MultiStringField(null=True)
     source_hash = models.CharField(max_length=32, db_index=True,
                                    editable=False)
@@ -452,6 +460,7 @@ class Unit(models.Model, base.TranslationUnit):
         user = kwargs.pop("user", self._log_user)
 
         self.pootle_path = self.store.pootle_path
+        self.language = self.store.translation_project.language
 
         if created:
             self._save_action = UNIT_ADDED