--- a/pootle/apps/pootle_store/models.py
+++ b/pootle/apps/pootle_store/models.py
@@ -312,8 +312,7 @@ class Unit(models.Model, base.TranslationUnit):
     # denormalized TP.language
     language = models.ForeignKey(
         "pootle_language.Language",
-        null=True,
-        blank=True,
+        null=False,
         db_index=True,
         related_name='units')