Denormalize project to unit to allow fast filtering and indexing together with language.
BEFORE: [version](https://github.com/translate/pootle/tree/2c525a484c1d81512639641adb784a095ae6f41c) before any get_units refactoring.
AFTER: current PR.
The timings are specifically for the get_units db call - with the uIds call switched off.

### /projects/translate
AFTER: ~~.004s~~ .002s
BEFORE: ∞ 

### /projects/firefox/translate
AFTER: ~~.004s~~ .002s
BEFORE: ~~154s~~, >5s

### /ru/firefox/translate
AFTER: ~~.29s~~ ~~.31s~~ .002s
BEFORE: 0.02s

### /ru/translate/
AFTER: ~~.29s~~ ~~.31s~~ .012s
BEFORE: 0.05s

### /projects/firefox/translate/browser/branding/official/brand.dtd.po
AFTER: ~~.26s~~ ~~.248s~~ .23s
BEFORE: 0.261s

### /ru/firefox/translate/browser/branding/official/brand.dtd.po
AFTER: ~~.001s~~ .001s
BEFORE: .001s

### /projects/firefox/translate/browser/branding/
AFTER: ~~.114s~~ ~~.104s~~ .099s
BEFORE: 0.2s

### /ru/firefox/translate/browser/branding/
AFTER: ~~137s~~ .001s
BEFORE: .001s