@julen 2016-03-03
After several rounds of testing here are the numbers I got, and while the site-wide and TP root query timings greatly improve, there looks to be quite a negative impact in the cross-language and cross-project views.

I was also testing locally on MySQL 5.7 first (it had [some gotchas when migrating](http://blog.wl0.org/2015/12/a-couple-of-mysql-5-7-gotchas-to-be-aware-of/)), although I had to give up because it seems the default configuration changed quite a bit and the overall test runs were taking forever.

<table>
<thead>
<tr>
<th><em>path</em></th>
<th><em>branch</em></th>
<th>MySQL 5.1 (stage)</th>
<th>MySQL 5.6 (local)</th>
</tr>
</thead>
<tbody>
<tr>
<td rowspan="3">/projects/translate/</td>
<td>en-release</td>
<td>74.364s (7q)</td>
<td>97.044s (8q)</td>
</tr>
<tr>
<td>master</td>
<td>129.148s (4q)</td>
<td>118.856s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>4.432s (74q)</td>
<td>4.255s (74q)</td>
</tr>
<tr>
<td rowspan="3">/projects/web_evernote/translate/</td>
<td>en-release</td>
<td>7.909s (5q)</td>
<td>11.551s (6q)</td>
</tr>
<tr>
<td>master</td>
<td>8.334s (5q)</td>
<td>8.764s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>57.837s (4q)</td>
<td>32.598s (5q)</td>
</tr>
<tr>
<td rowspan="3">/ru/web_evernote/translate/</td>
<td>en-release</td>
<td>0.378s (5q)</td>
<td>0.218s (6q)</td>
</tr>
<tr>
<td>master</td>
<td>0.501s (4q)</td>
<td>0.37s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>0.062s (4q)</td>
<td>0.053s (5q)</td>
</tr>
<tr>
<td rowspan="3">/ru/translate/</td>
<td>en-release</td>
<td>2.504s (7q)</td>
<td>4.062s (8q)</td>
</tr>
<tr>
<td>master</td>
<td>4.435s (4q)</td>
<td>2.91s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>19.941s (4q)</td>
<td>9.712s (5q)</td>
</tr>
<tr>
<td rowspan="3">/projects/web_evernote/translate/<br/>gwt/com/evernote/web/gwt/<br/>client/v2/i18n/<br/>FocusMessages.properties.po</td>
<td>en-release</td>
<td>0.404s (5q)</td>
<td>0.253s (6q)</td>
</tr>
<tr>
<td>master</td>
<td>0.609s (4q)</td>
<td>0.521s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>16.664s (4q)</td>
<td>10.148s (5q)</td>
</tr>
<tr>
<td rowspan="3">/ru/web_evernote/translate/<br/>gwt/com/evernote/web/gwt/<br/>client/v2/i18n/<br/>FocusMessages.properties.po</td>
<td>en-release</td>
<td>0.014s (5q)</td>
<td>0.009s (6q)</td>
</tr>
<tr>
<td>master</td>
<td>0.022s (4q)</td>
<td>0.006s (5q)</td>
</tr>
<tr>
<td>PR</td>
<td>0.129s (4q)</td>
<td>0.121s (5q)</td>
</tr>
<tr>
<td rowspan="3">/projects/web_evernote/translate/<br/>gwt/com/evernote/web/<br/>gwt/client/v2/i18n/</td>
<td>en-release</td>
<td>0.851s (6q)</td>
<td>0.618s (7q)</td>
</tr>
<tr>
<td>master</td>
<td>1.399s (5q)</td>
<td>0.945s (6q)</td>
</tr>
<tr>
<td>PR</td>
<td>23.601s (4q)</td>
<td>13.297s (6q)</td>
</tr>
<tr>
<td rowspan="3">/ru/web_evernote/translate/<br/>gwt/com/evernote/web/<br/>gwt/client/v2/i18n/</td>
<td>en-release</td>
<td>0.025s (6q)</td>
<td>0.016s (7q)</td>
</tr>
<tr>
<td>master</td>
<td>0.06s (5q)</td>
<td>0.017s (6q)</td>
</tr>
<tr>
<td>PR</td>
<td>0.017s (4q)</td>
<td>0.041s (6q)</td>
</tr>
</tbody>
</table>