@iafan 2016-03-09
I know that you guys have put a good deal of time and energy into this PR, but I personally have mixed feelings about this approach. I absolutely agree that we need to improve the performance when working with large installations, but this particular way of addressing the problem seems like a brute-force one: it increases the code complexity and DB size while not giving any reasonable reproducible speedups that everybody would agree with. The most frustrating thing would be for all of us to rush this into production, do the costly migrations only to realize several months later that there's a better approach and we need to undo the code and migrations and implement this in a more scalable fashion. 

I would suggest us to test other approaches before sticking with this PR. For example, we now use ElasticSearch for TM lookups, and it proved to be a great solution in this particular area. Did you consider using ES for indexing and retrieving units as well? I have a feeling that it has a potential of  drastically improving the search speed in all scenarios, without complicating the main codebase much. I bet writing a script that just populates ES index from Pootle units will take an hour or so to implement, and then we could all experiment with it, and run queries that return needed pages of data from the entire data set sorted by various fields (store path + unit index, creation date, etc.). I would strongly suggest doing at least this experiment before deciding what to do with this PR. I might be wrong in my assumptions, but if I am, this will only give extra credibility to this PR and will make all of us feel more easy about pushing it into production.

Again, I don't want to insult anyone with this comment, and you @phlax specifically do a really amazing work in pushing the project forward. I just want to make sure everybody who works on the project agrees on the important architectural decisions, fully understands why they are necessary and what pros/cons these changes bring.