@phlax 2016-03-09
hi @iafan thanks for taking the time to comment

I would like to make a few points in respect to this

Current problems irrespective of performance
---

- the current code doesnt order correctly ie it groups units by (pootle_path, index) that havent been ordered that way in the total dataset - leading to various problems/inconsistencies
- some of the ordering is also not what you might expect - in particular oldest/newest is neither the earliest/most recent change/creation time - but when unit was translated (mostly)

Current expectation/desired functionality
---

 - imho and afaict most peoples expectations we *should* order on pootle_path,index as this is basically by_store, by_order in store
- we also expect (for good reason) ordering by other metrics - ie mtime, submitted_time, priority etc


Performance issues
---

- current setup requires complex sorts on not only very large Unit tables, but very large unit sets within that table 
- this is causing *massive* performance issues for us before/after this PR - but defo less so with this PR in place if we expect proper pootle_path,index ordering other issues also need to be resolved to really fix the problem).
 
Performance fixes
---

- Denormalizing fields where bottlenecks occur is good practice
- Adding multi-column indeces for complex/common sorts/filters with MySQL is necessary for large tables and good practice
- Getting the optimal/*minimal* denormalization is defo good practice and im working really hard to try and achieve that
- The performance issues that we are seeing on large datasets are going to be reflected in smaller installs of Pootle also so my feeling is that bad performance is highlighting to us deeper issues rather than just some bad performance on specific TPs on MLO/EN

Using Elasticsearch or other Indexers
---

- Using a (distributed/~async) indexer bring its own complexity - particularly where Units are being edited in quick succession - so this requires some testing
- Making ElasticSearch/OtherIndexer a hard dependency is less conducive to someone installing Pootle easily so i would really rather avoid this if possible (for now at least)
- Local TM is currently not a *hard*-dep so while its useful that we already have those docs/install path etc its not the same as forcing everyone who installs Pootle to install ElasticSearch/OtherIndexer
- Optimizing Indexes/Facets within WhateverIndexers is going to take some time (like lots of very disk IO-intensive time)
- For the record I absolutely see this as the way forward for large Pootle installations in case that is not clear already